+++ packages/ccs-frontend/tasks/index.mjs
@@ -2,6 +2,7 @@
  * Build tasks
  */
 export { assets } from './assets.mjs'
+export { compile as declarations } from './declarations.mjs'
 export { compile as fixtures } from './fixtures.mjs'
 export { compile as scripts } from './scripts.mjs'
 export { compile as styles } from './styles.mjs'