+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The following CCS helpers have been added:
   - Contact us
 
+- Add a type declaration files
+
 ## [1.0.0] - 2024-06-07
 
 ### Added